pyproject.toml
[project]
name = "chatgpt-micro-cap-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib==3.8.4",
    "numpy==2.3.2",
    "pandas==2.2.2",
    "pydoll-python>=2.8.0",
    "yfinance==0.2.65",
]
